Déplacement des fonctions d'insertion la ou elles sont utilisées.
