+++ core/requetes.php
@@ -334,36 +334,6 @@ function nb_categories_poubelles(PDO $bdd): int {
   return (int) $nombre_categories;
 }
 
-// Insere une collecte dans la base.
-// On ecrit le nombre de categories maxi dans la varialble $nombreCategories
-// HACK C'est pas parfait il faudrait comparer aux differents ID de la base.
-// Si des utilisateurs suppriment des objet sa la main c'est la galere...
-// Mais d'un cote niveau tracabilite ils devraient pas supprimer de categories...
-// genere une exception si les masses sont inferieurs a 0.
-
-function insert_items_collecte(PDO $bdd, int $id_collecte, $collecte, $items) {
-  $nombreCategories = nb_categories_dechets_item($bdd);
-  $req = $bdd->prepare('INSERT INTO pesees_collectes
-                            (timestamp, masse, id_collecte, id_type_dechet, id_createur, id_last_hero)
-                        VALUES (:timestamp, :masse, :id_collecte, :id_type_dechet, :id_createur, :id_createur1)');
-  $req->bindValue(':timestamp', $collecte['timestamp']->format('Y-m-d H:i:s'), PDO::PARAM_STR);
-  $req->bindValue(':id_collecte', $id_collecte, PDO::PARAM_INT);
-  $req->bindValue(':id_createur', $collecte['id_user'], PDO::PARAM_INT);
-  $req->bindValue(':id_createur1', $collecte['id_user'], PDO::PARAM_INT);
-  foreach ($items as $item) {
-    $masse = (float) parseFloat($item['masse']);
-    $type_dechet = (int) parseInt($item['type']);
-    if ($masse > 0.00 && $type_dechet <= $nombreCategories) {
-      $req->bindValue(':masse', $masse);
-      $req->bindValue(':id_type_dechet', $type_dechet, PDO::PARAM_INT);
-      $req->execute();
-    } else {
-      $req->closeCursor();
-      throw new UnexpectedValueException('masse <= 0.0 ou type item inconnu');
-    }
-  }
-}
-
 // TODO: Refactorer ça trop d'info
 function pesee_collectes_id(PDO $bdd, int $id_collecte): int {
   $sql = 'SELECT
@@ -403,137 +373,6 @@ function collecte_id(PDO $bdd, int $id): array {
 function moyens_paiements(PDO $bdd): array {
   $sql = 'SELECT id, nom, couleur, visible, description, timestamp FROM moyens_paiement';
   return fetch_all($sql, $bdd);
-}
-
-function insert_collecte(PDO $bdd, array $collecte): int {
-  $req = $bdd->prepare('INSERT INTO collectes
-                            (timestamp, id_type_collecte,
-                            localisation, id_point_collecte,
-                            commentaire, id_createur, id_last_hero)
-                          VALUES (:timestamp, :id_type_action,
-                                  :localite, :id_point,
-                                  :commentaire, :id_createur, :id_createur1)');
-  $req->bindValue(':timestamp', $collecte['timestamp']->format('Y-m-d H:i:s'), PDO::PARAM_STR);
-  $req->bindValue(':id_type_action', $collecte['id_type_action'], PDO::PARAM_INT);
-  // HACK: virer les adherants de la base ou juste changer le type en booleen serieusement...
-  $req->bindValue(':localite', $collecte['localite'], PDO::PARAM_INT);
-  $req->bindValue(':id_point', $collecte['id_point'], PDO::PARAM_INT);
-  $req->bindValue(':commentaire', $bdd->quote($collecte['commentaire']), PDO::PARAM_STR);
-  $req->bindValue(':id_createur', $collecte['id_user'], PDO::PARAM_INT);
-  $req->bindValue(':id_createur1', $collecte['id_user'], PDO::PARAM_INT);
-  $req->execute();
-  return (int) $bdd->lastInsertId();
-}
-
-function insert_items_sorties(PDO $bdd, int $id_sorties, $sortie, $items) {
-  $nombreCategories = nb_categories_dechets_item($bdd);
-  $req = $bdd->prepare('INSERT INTO pesees_sorties (timestamp, masse,  id_sortie, id_type_dechet, id_createur, id_last_hero)
-                            VALUES(:timestamp, :masse, :id_sortie, :id_type_dechet, :id_createur, :id_createur1)');
-  $req->bindValue(':timestamp', $sortie['timestamp']->format('Y-m-d H:i:s'), PDO::PARAM_STR);
-  $req->bindValue(':id_sortie', $id_sorties, PDO::PARAM_INT);
-  $req->bindValue(':id_createur', $sortie['id_user'], PDO::PARAM_INT);
-  $req->bindValue(':id_createur1', $sortie['id_user'], PDO::PARAM_INT);
-  foreach ($items as $item) {
-    $masse = (float) parseFloat($item['masse']);
-    $type_dechet = (int) parseInt($item['type']);
-    if ($masse > 0.00 && $type_dechet <= $nombreCategories) {
-      $req->bindValue(':masse', $masse);
-      $req->bindValue(':id_type_dechet', $type_dechet, PDO::PARAM_INT);
-      $req->execute();
-    } else {
-      $req->closeCursor();
-      throw new UnexpectedValueException('masse <= 0.0 ou type item inconnu');
-    }
-  }
-}
-
-function insert_evac_sorties(PDO $bdd, int $id_sorties, $sortie, $items) {
-  $nombreCategories = nb_categories_dechets_evac($bdd);
-  $req = $bdd->prepare('INSERT INTO pesees_sorties (timestamp, masse,  id_sortie, id_type_dechet_evac, id_createur)
-                            VALUES(:timestamp, :masse, :id_sortie, :id_type_dechet_evac, :id_createur)');
-  $req->bindValue(':timestamp', $sortie['timestamp']->format('Y-m-d H:i:s'), PDO::PARAM_STR);
-  $req->bindValue(':id_sortie', $id_sorties, PDO::PARAM_INT);
-  $req->bindValue(':id_createur', $sortie['id_user'], PDO::PARAM_INT);
-  foreach ($items as $item) {
-    $masse = (float) parseFloat($item['masse']);
-    $type_dechet = (int) parseInt($item['type']);
-    if ($masse > 0.00 && $type_dechet <= $nombreCategories) {
-      $req->bindValue(':masse', $masse);
-      $req->bindValue(':id_type_dechet_evac', $type_dechet, PDO::PARAM_INT);
-      $req->execute();
-    } else {
-      $req->closeCursor();
-      throw new UnexpectedValueException('masse <= 0.0 ou type item inconnu');
-    }
-  }
-}
-
-function insert_poubelle_sorties(PDO $bdd, int $id_sorties, $sortie, $items) {
-  $nombreCategories = nb_categories_poubelles($bdd);
-  $req = $bdd->prepare('INSERT INTO pesees_sorties (timestamp, masse, id_sortie, id_type_poubelle, id_createur)
-                            VALUES(:timestamp, :masse, :id_sortie, :id_type_poubelle, :id_createur)');
-  $req->bindValue(':timestamp', $sortie['timestamp']->format('Y-m-d H:i:s'), PDO::PARAM_STR);
-  $req->bindValue(':id_sortie', $id_sorties, PDO::PARAM_INT);
-  $req->bindValue(':id_createur', $sortie['id_user'], PDO::PARAM_INT);
-  foreach ($items as $item) {
-    $masse = (float) parseFloat($item['masse']);
-    $type_dechet = (int) parseInt($item['type']);
-    if ($masse > 0.00 && $type_dechet <= $nombreCategories) {
-      $req->bindValue(':masse', $masse);
-      $req->bindValue(':id_type_poubelle', $type_dechet, PDO::PARAM_INT);
-      $req->execute();
-    } else {
-      $req->closeCursor();
-      throw new UnexpectedValueException('masse <= 0.0 ou type item inconnu');
-    }
-  }
-}
-
-function specialise_sortie(PDOStatement $stmt, $sortie) {
-  $classe = $sortie['classe'];
-  // Sorties Dons
-  if ($classe === 'sorties') {
-    $stmt->bindvalue(':type_sortie', $sortie['type_sortie'], PDO::PARAM_INT);
-    $stmt->bindvalue(':id_filiere', 0, PDO::PARAM_INT);
-    $stmt->bindvalue(':id_convention', 0, PDO::PARAM_INT);
-    // Sorties recycleur
-  } elseif ($classe === 'sortiesr') {
-    $stmt->bindvalue(':type_sortie', 0, PDO::PARAM_INT);
-    $stmt->bindvalue(':id_filiere', $sortie['type_sortie'], PDO::PARAM_INT);
-    $stmt->bindvalue(':id_convention', 0, PDO::PARAM_INT);
-    // Sorties conventions
-  } elseif ($classe === 'sortiesc') {
-    $stmt->bindvalue(':type_sortie', 0, PDO::PARAM_INT);
-    $stmt->bindvalue(':id_filiere', 0, PDO::PARAM_INT);
-    $stmt->bindvalue(':id_convention', $sortie['type_sortie'], PDO::PARAM_INT);
-  } elseif ($classe === 'sortiesp' || $classe === 'sortiesd') {
-    $stmt->bindvalue(':type_sortie', 0, PDO::PARAM_INT);
-    $stmt->bindvalue(':id_filiere', 0, PDO::PARAM_INT);
-    $stmt->bindvalue(':id_convention', 0, PDO::PARAM_INT);
-  } else {
-    throw new UnexpectedValueException('class de sortie inconnue');
-  }
-  return $stmt;
-}
-
-function insert_sortie(PDO $bdd, array $sortie): int {
-  $sql = 'INSERT INTO sorties (timestamp, id_filiere, id_convention,
-                                id_type_sortie, classe,
-                                id_point_sortie, commentaire, id_createur)
-          VALUES(:timestamp, :id_filiere, :id_convention,
-                 :type_sortie, :classe,
-                 :id_point_sortie, :commentaire, :id_createur)  ';
-  $req = $bdd->prepare($sql);
-  $req->bindvalue(':timestamp', $sortie['timestamp']->format('Y-m-d H:i:s'), PDO::PARAM_STR);
-  $req->bindvalue(':classe', $sortie['classe'], PDO::PARAM_STR);
-
-  $req = specialise_sortie($req, $sortie);
-
-  $req->bindvalue(':id_point_sortie', $sortie['id_point_sortie'], PDO::PARAM_INT);
-  $req->bindvalue(':commentaire', $bdd->quote($sortie['commentaire']), PDO::PARAM_STR);
-  $req->bindvalue(':id_createur', $sortie['id_user'], PDO::PARAM_INT);
-  $req->execute();
-  return $bdd->lastInsertId();
 }
 
 function structure(PDO $bdd): array {
